% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{extraire.nir}
\alias{extraire.nir}
\title{Traitement du NIR (numéro d'inscription au répertoire des personnes physiques).}
\usage{
extraire.nir(Base, année)
}
\arguments{
\item{Base}{data.table contenant au moins une variable nommée Nir décrivant le NIR.}

\item{année}{Année civile à la fin de laquelle est évalué l'âge de l'individu.}
}
\value{
Une base data.table ayant la forme suivante (les bornes d'âge ne sont pas impératives) :
       \tabular{ccc}{
         age \tab Hommes \tab Femmes \cr
         15  \tab   0  \tab    1   \cr
         16  \tab   NA \tab    2   \cr
         17  \tab   1  \tab    3   \cr
         18  \tab  409 \tab    52  \cr
         ... \tab  ... \tab ...    \cr
         68  \tab 2216 \tab    NA
       }
}
\description{
Extrait la répartition par âge et sexe des individus ayant un NIR.
}
\examples{
extraire.nir(Base, 2012)
}
